# [jvm-packages] Switch Task mode to OpenMP mode after data conversion