--- a/include/xgboost/c_api.h
+++ b/include/xgboost/c_api.h
@@ -364,6 +364,20 @@ XGB_DLL int XGDMatrixSliceDMatrixEx(DMatrixHandle handle,
                                     bst_ulong len,
                                     DMatrixHandle *out,
                                     int allow_groups);
+
+
+/*!
+ * \brief combine right dmatrix into left, return the left
+ * \param handle_left instance of data matrix to be merged into
+ * \param handle right instance of data matrix to be merged
+ * \param out the left instance of data matrix
+ * \return 0 when success, -1 when failure happens
+ */
+XGB_DLL int XGDMatrixCombineDMatrix(DMatrixHandle handle_left,
+                                  DMatrixHandle handle_right,
+                                  uint64_t total_size,
+                                  DMatrixHandle* out);
+
 /*!
  * \brief free space in data matrix
  * \return 0 when success, -1 when failure happens