--- a/include/xgboost/data.h
+++ b/include/xgboost/data.h
@@ -187,8 +187,7 @@ struct Entry {
   /*! \brief feature value */
   bst_float fvalue;
   /*! \brief default constructor */
-  Entry() {
-  }
+  Entry() = default;
   /*!
    * \brief constructor with index and value
    * \param index The feature or row index.