One big issue of current xgboost training on Spark is that ETL usually run in task mode which sets spark.task.cpus = 1. But xgboost training is more efficient on openMP mode which sets spark.task.cpus = spark.executor.cores = xgbparam.nthread. We usually save the ETL result into a parquet file, then start a new spark context with task.cpus=nthreads, then we load the data, convert into DMatrix, start training. In this way the conversion has to run in little threads and takes very long time. e.g. a conversion of 150M Row x 46 columns takes about 870s on a 2x28 config, takes ~250s on a 7x8 config.  Even worse, the inference is more efficient on task mode, so we have to save the model and start another new context to do inference.

It's really not a good solution. The patch gives a way to swich task mode to OMP mode AFTER data conversion, then the training run in OMP mode, then the inference goes back to task mode. It solved the model swtich issue, accelerately the data conversion process greatly and be able to use one or two executor per node which is more efficient for training.

The basic logic is as below:
    1. repartition input data set to N x executor#. N is integer.
    2. Create DMatrix in each task thread of executor
    3. Force caching the Watches[DMatrix*] in executor's block manager
    4. Use a customized coalescePartitioner to combine the Watches[DMatrix*] into a large DMatrix. Set preferred location to the executor
    5. Train xgboost from the large DMatrix in task thread of an executor

Because we cached the pointer of DMatrix in block manager instead of the converted data itself, so we must make sure the training stage must be scheduled to the same executor. we can set spark.locality.wait.process to a large number for this. Spark3.0's stage level resource management can change task.cpus at run time but it doesn't solve the issue automatically. We still need to combine the converted DMatrix of each executor into a large one. It either can't solve the data locality issue, we still need to set wait time a large number.

Here is the result, with 7x8 configuration, the patch can decrease the data conversion time from 250s to 70s. with 2x28 configuration, it can decrease from 870s to 160s. It's possible to further optimize later.

![image](https://user-images.githubusercontent.com/47296334/84233478-427d7500-aaa7-11ea-9e9d-75592f06f6f4.png)

It only support trainForNonRanking and evalSetsMap.isEmpty now. I will add other 3 cases.