@FelixYBW 2020-07-20
Added Ranking support.