@zhangzhang10 2020-08-28
@hcho3 Hi can you please help take a look at PR5774? One of the CI checks (jenkins pr-merge) failed due to "time out" error. How shall we fix such a problem? Thanks!