@hcho3 2020-08-28
@zhangzhang10 It probably is due to hitting [the daily budget limit in the CI server](https://xgboost-ci.net/dashboard/). I restarted the tests.