@hcho3 2020-08-28
@zhangzhang10 Actually it appears that this unit test in the JVM package is hanging:
```
[2020-08-28T17:30:10.151Z] Tracker started, with env={DMLC_NUM_SERVER=0, DMLC_TRACKER_URI=172.17.0.3, DMLC_TRACKER_PORT=9091, DMLC_NUM_WORKER=4}
[2020-08-28T17:30:10.151Z]  xgbtck executorcoalesce Some(executor_d7f5674fd1f3_driver) : 0 1 2 3
```
We need to fix this test for this PR to be merged.