@hcho3 2020-08-28
This is the test that's hanging:
https://github.com/dmlc/xgboost/blob/d7b45fbcafcce8295a94858a996482fdb9091f3e/jvm-packages/xgboost4j-spark/src/test/scala/ml/dmlc/xgboost4j/scala/spark/ParameterSuite.scala#L26-L41