@zhangzhang10 2020-08-28
@hcho3 Thank you for catching the failed test. I'll fix it!