@CodingCat 2020-08-30
@FelixYBW thanks for the contribution, but I don't think this is a solution to handle this problem and this should be a problem handled in XGBoost side. My concerns come from 

1. "is saving transformed data to Parquet really a bad idea"?

When I started this project in school, I also thought eliminating the intermediate data is good. However, in prod case, training of XGB is "much more fragile" than ETL processing itself. Spot instance, preemption and various hw faults can take down anyone of your workers, and XGBoost training will just crash. If you couple ETL and training together, that means, you need to redo ETL in the case of a failed training .


2. IIUC, the key thing to make your method work is to schedule task to a particular executor saving DMatrix pointer. To make this happen, you are setting `spark.locality.wait.process to a large number for this`

first of all, even you set spark.locality.wait.process  to a large number, there is no guarantee that your task will eventually get scheduled to that executor. 

second, setting spark.locality.wait.process to a super large value will hurt performance of ETL in storage-compute disaggregated architecture (which is the trend in industry). In such an architecture, we usually set spark.locality.wait to 0s to avoid wasting of compute resources for pursuing locality which doesn't exist at all.

3. I do not think as a ML framework, XGBoost/XGBoost-Spark should tackle this problem. The underlying framework like Spark should consider per-stage resource claim , just like they should consider only allocating GPU to a particular stage of Spark applications instead of letting XGBoost-Spark to handle this for them.

BTW, I don't think Spark 3.0 has been able to set per-stage cpu core, the PR hasn't been merged (I might be wrong)