@zhangzhang10 2020-09-08
@CodingCat, you may have misunderstood this PR. Although @FelixYBW 's original comment mentioned ETL, the main point of this PR isn't about ETL or about doing away intermediate data. The main point is to provide improved overall training performance when OpenMP optimization is enabled. On Spark, OpenMP optimization is enabled by allocating multiple threads for each training task (controlled by parameter `nthread`). But the step before training, i.e. constructing DMatrix data structure from input data is still slow because it is not as well parallelized. With this PR, we can make the DMatrix construction step run faster by using a larger number of single-threaded tasks. Then, the DMatrix data structures created by tasks on the same executor are combined. The training step then takes advantage of OpenMP by using multi-threaded tasks, albeit a fewer number of them. 

From a resource management point of view, this PR, while maintaining the same amount of resource (i.e. compute cores), automatically shifts from using many single-threaded tasks to create DMatrix, to using fewer multi-threaded task for training. This way, it aims to provide a better overall training experience.