@CodingCat 2020-09-09
> But the step before training, i.e. constructing DMatrix data structure from input data is still slow because it is not as well parallelized. With this PR, we can make the DMatrix construction step run faster by using a larger number of single-threaded tasks. Then, the DMatrix data structures created by tasks on the same executor are combined. The training step then takes advantage of OpenMP by using multi-threaded tasks, albeit a fewer number of them.

sure, but as I said, "even you set spark.locality.wait.process to a large number, there is no guarantee that your task will eventually get scheduled to that executor." and `setting spark.locality.wait.process to a super large value will hurt performance of ETL in storage-compute disaggregated architecture (which is the trend in industry).`



from the description `a conversion of 150M Row x 46 columns takes about 870s on a 2x28 config, ` training 150M on a single task seems pretty heavy, why not assigning more workers?