@zhangzhang10 2020-10-01
> from the description `a conversion of 150M Row x 46 columns takes about 870s on a 2x28 config, ` training 150M on a single task seems pretty heavy, why not assigning more workers?

Let me answer this question first. The reason for not assigning more workers is to keep a lower memory overhead, so more memory can be given to the training tasks. Consequently, we can train a larger dataset this way. The advantage can be significant. We carefully benchmarked memory consumptions using a random dataset with 50 features to see how big a dataset we can train. The result is shown below (the experiments were done on an Intel Xeon 8180 node with 56 cores and 380GB memory):

|Config (num_workers x cores_per_worker)   |Max data size (num_rows)   |Peak mem (GB)   |
|:-:|:-:|:-:|
|2x28   |550 million   |360   |
|7x8   |300 million   |300   |

Using the 2x28 config we can train a dataset 1.8x larger than what we can train using the 7x8 config. The incentive for using fewer workers is obvious. However, without this PR the 2x28 config means a very long data conversion time, which diminishes its benefit. With this PR the data conversion time is substantially shortened so that users can easily reap the benefit.