@carsonwang 2020-10-09
> sure, but as I said, "even you set spark.locality.wait.process to a large number, there is no guarantee that your task will eventually get scheduled to that executor." and `setting spark.locality.wait.process to a super large value will hurt performance of ETL in storage-compute disaggregated architecture (which is the trend in industry).`

@CodingCat , the tasks should be scheduled correctly even if the locality wait is 0.  Setting spark.locality.wait.process to 0 means the [getAllowedLocalityLevel](https://github.com/apache/spark/blob/0811666ab104b41cf189233439f4158b18bc8282/core/src/main/scala/org/apache/spark/scheduler/TaskSetManager.scala#L434) will always be ANY. But Spark schedules the tasks in [increasing order of locality levels](https://github.com/apache/spark/blob/0811666ab104b41cf189233439f4158b18bc8282/core/src/main/scala/org/apache/spark/scheduler/TaskSchedulerImpl.scala#L579) which is controlled by `maxLocality` and it will start from PROCESS_LOCAL. For example, if we have N executors and each executor has M cores. In the training stage, there are only N tasks and every task has a preferred location (executor). In this case, Spark will first schedule tasks using maxLocality = PROCESS_LOCAL. Even though  spark.locality.wait.process is 0, all N tasks will be scheduled to the preferred location correctly with the maxLocality constraint. Please correct if anything was missing here.